delivery-kid/pinning-service/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async HTTP client
httpx>=0.26.0

# Wallet signature verification
eth-account>=0.11.0

# SSE for progress streaming
sse-starlette>=1.8.0

# Settings management
pydantic-settings>=2.1.0

# Audio processing
pydub>=0.25.1

# YAML parsing (release enrichment)
pyyaml>=6.0
